Re-introduced 1-class-per-file linter
